+++ include/shell/commands.h
@@ -40,6 +40,7 @@ void cmd_echo(const char* args);
 void cmd_secho(const char* args);
 void cmd_memstat(const char* args);
 void cmd_heapstat(const char* args);
+void cmd_int(const char* args);
 void cmd_grep(const char* args);
 
 // fs