+++ kernel/shell/commands.c
@@ -28,6 +28,7 @@ ShellCommand command_table[] = {
     {"secho", cmd_secho, "Write text to the serial port (COM1)"},
     {"memstat", cmd_memstat, "Memory statistics"},
     {"heapstat", cmd_heapstat, "Verify heap health"},
+    {"int", cmd_int, "Jump to given interrupt"},
     {"grep", cmd_grep, "Searches text for matching patterns"},
     {"cd", cmd_cd, "Change directory"},
     {"cat", cmd_cat, "Read file"},